+++ DatadogTrace/Sources/Integrations/TracingURLSessionHandler.swift
@@ -173,7 +173,8 @@ internal struct TracingURLSessionHandler: DatadogURLSessionHandler {
             interception.isFirstPartyRequest, // `Span` should be only send for 1st party requests
             interception.origin != "rum", // if that request was tracked as RUM resource, the RUM backend will create the span on our behalf
             let tracer = tracer,
-            let resourceMetrics = interception.metrics,
+            let startTime = interception.fetchStartDate,
+            let endTime = interception.fetchEndDate,
             let resourceCompletion = interception.completion
         else {
             return
@@ -220,7 +221,7 @@ internal struct TracingURLSessionHandler: DatadogURLSessionHandler {
             span = tracer.startSpan(
                 spanContext: context,
                 operationName: "urlsession.request",
-                startTime: resourceMetrics.fetch.start
+                startTime: startTime
             )
         } else if Sampler(samplingRate: samplingRate).sample() {
             // Span context may not be injected on iOS13+ if `URLSession.dataTask(...)` for `URL`
@@ -242,7 +243,7 @@ internal struct TracingURLSessionHandler: DatadogURLSessionHandler {
             span = tracer.startSpan(
                 spanContext: context,
                 operationName: "urlsession.request",
-                startTime: resourceMetrics.fetch.start
+                startTime: startTime
             )
         } else {
             return
@@ -278,16 +279,16 @@ internal struct TracingURLSessionHandler: DatadogURLSessionHandler {
         }
 
         if let history = contextReceiver.context.applicationStateHistory {
-            let fetchDuration = resourceMetrics.fetch.start...resourceMetrics.fetch.end
+            let fetchDuration = startTime...endTime
             let foregroundDuration = history.foregroundDuration(during: fetchDuration)
             span.setTag(key: SpanTags.foregroundDuration, value: foregroundDuration.dd.toNanoseconds)
 
-            let didStartInBackground = history.state(at: resourceMetrics.fetch.start) == .background
-            let doesEndInBackground = history.state(at: resourceMetrics.fetch.end) == .background
+            let didStartInBackground = history.state(at: startTime) == .background
+            let doesEndInBackground = history.state(at: endTime) == .background
             span.setTag(key: SpanTags.isBackground, value: didStartInBackground || doesEndInBackground)
         }
 
-        span.finish(at: resourceMetrics.fetch.end)
+        span.finish(at: endTime)
     }
 
     /// Creates a helper struct with collected elements from a possible parent span context.